Add memcheck to build script and CI test
