+++ .ci/docker/env.list
@@ -11,3 +11,4 @@ BUILD_DOCS
 CODECOV
 SUDO
 BUILD_DARTPY
+MEMCHECK
